ai-model/requirements.txt
Flask
flask-cors
Pillow
numpy
scikit-learn